Make retry on ratelimit configurable
